Get available exchanges

Get a vector of all available exchanges passed into the source argument of the get-functions.

Usage

availableExchanges(type = "ohlc")

Arguments

type

character-vector of length 1. See details

Value

An invisible()

character vector containing available exchanges

Details

Available types

ohlc: Open, High, Low, Close and Volume

lsratio: Long-Short ratio

fundingrate: Funding rates

interest: Open perpetual contracts on both sides

Limits

The endpoints supported by the available_exchanges() are not uniform, so exchanges available for, say, get_lsratio() is not necessarily the same as those available for get_quote()

See also

Author

Serkan Korkmaz

Examples

# script:
# date: 2023-10-06
# author: Serkan Korkmaz, serkor1@duck.com
# objective:
# script start;

## return all
## available exchanges
cryptoQuotes::available_exchanges()
#> ℹ Available Exchanges:
#> ✔ binance, bitmart, bybit, kraken, kucoin

# script end;