Get available intervals

Get available intervals for the available_tickers() on the available_exchanges().

Usage

available_intervals(
   source = "binance",
   type   = "ohlc",
   futures = TRUE
)

Arguments

source

A character-vector of length 1. binance by default. See available_exchanges() for available exchanges.

type

character-vector of length 1. One of,

"ohlc" - Available exchanges for Open, High, Low, Close and Volume market data. See the get_quote()-function.

"lsratio" - Available exchanges for Long-Short ratios. See the get_lsratio()-function.

"fundingrate" - Available exchanges for Funding rates. See the get_fundingrate()-function.

"interest" - Available exchanges for Open interest on perpetual contracts on both sides. See the get_openinterest()-function.

futures

A logical-vector of length 1. TRUE by default. Returns futures market if TRUE, spot market otherwise.

Value

An invisible()

character-vector containing the available intervals on the exchange, market and endpoint.

Sample output

#> i Available Intervals at "bybit" (futures):
#> v 1m, 3m, 5m, 15m, 30m, 1h, 2h, 4h, 6h, 12h, 1d, 1M, 1w
#> [1] "1m"  "3m"  "5m"  "15m" "30m" "1h"

Details

The endpoints supported by the available_exchanges() are not uniform, so exchanges available for, say, get_lsratio() is not necessarily the same as those available for get_quote()

See also

Author

Serkan Korkmaz

Examples

if (FALSE) {
  # script start;

  # available intervals
  # at kucoin futures market
  cryptoQuotes::available_intervals(
    source  = 'kucoin',
    futures = TRUE,
    type    = "ohlc"
  )

  # available intervals
  # at kraken spot market
  cryptoQuotes::available_intervals(
    source  = 'kraken',
    futures = FALSE,
    type    = "ohlc"
  )

  # script end;
}