cryptoQuotes: Open Access to Cryptocurrency Market Data, Sentiment Indicators and Interactive Charts

This high-level API client provides open access to cryptocurrency market data, sentiment indicators, and interactive charting tools. The data is sourced from major cryptocurrency exchanges via 'curl' and returned in 'xts'-format. The data comes in open, high, low, and close (OHLC) format with flexible granularity, ranging from seconds to months. This flexibility makes it ideal for developing and backtesting trading strategies or conducting detailed market analysis.

See also

Author

Maintainer: Serkan Korkmaz serkor1@duck.com (ORCID) [contributor, copyright holder]

Other contributors:

Jonas Cuzulan Hirani jjh@vive.dk (ORCID) [contributor]