Add Volume-Weighted Moving Average (VWAP) to the chart

A high-level plotly::add_lines()-wrapper function that interacts with {TTR}'s moving average family of functions. The function adds moving average indicators to the main chart().

Usage

vwap(
 price  = "close",
 n      = 10,
 ratio = NULL,
 ...
)

Arguments

price

A character-vector of length 1. "close" by default. The name of the vector to passed into TTR::VWAP()

n

Number of periods to average over. Must be between 1 and nrow(x), inclusive.

ratio

A smoothing/decay ratio. ratio overrides wilder in EMA.

...

For internal use. Please ignore.

Value

A plotly::plot_ly()-object

See also

Author

Serkan Korkmaz

Examples

# script start;

cryptoQuotes::chart(
  ticker = BTC,
  main = kline(),
  indicator = list(
    cryptoQuotes::ema(n = 7),
    cryptoQuotes::sma(n = 14),
    cryptoQuotes::wma(n = 21)
  )
)

# script end;