Read and Write xts-objects

The write_xts()- and read_xts()-functions are zoo::write.zoo()- and zoo::read.zoo()-wrapper functions.

Usage

# write XTS-object
write_xts(
 x,
 file,
 ...
)

# read XTS-object
read_xts(
file
)

Arguments

x

An <xts>-object.

file

character string or strings giving the name of the file(s) which the data are to be read from/written to. See read.table and write.table for more information. Alternatively, in read.zoo, file can be a connection or a data.frame (e.g., resulting from a previous read.table call) that is subsequently processed to a "zoo" series.

...

further arguments passed to other functions. In the read.*.zoo the arguments are passed to the function specified in read (unless file is a data.frame already). In write.zoo the arguments are passed to write.table.

Details

When reading and writing <xts>-objects the attributes does not follow the object.

See also

Author

Serkan Korkmaz