Get a quote on a cryptopair from one of the supported exchanges

Open, High, Low, Close, and Volume (OHLCV) quotes are essential pieces of information used to analyze the price and trading activity of a financial asset over a specific time frame.

Usage

getQuote(
 ticker,
 source   = 'binance',
 futures  = TRUE,
 interval = '1d',
 from     = NULL,
 to       = NULL
)

Arguments

ticker

A character vector of length 1. Uppercase. See availableTickers() for available tickers.

source

A character vector of length 1. See availableExchanges() for available exchanges.

futures

A logical value. Returns futures market if TRUE, spot market otherwise.

interval

A character vector of length 1. See availableIntervals() for available intervals.

from

An optional character vector of length 1. Given in %Y-%m-%d format.

to

An optional character vector of length 1. Given in %Y-%m-%d format.

Value

an xts object with Open, High, Low, Close and Volume. If futures = TRUE the prices are last prices.

Details

If only from is provided 100 pips are returned up to Sys.Date().

If only to is provided 100 pips up to the specified date is returned.

If from and to are both NULL 100 pips returned up to Sys.Date()

Examples

# 1) Load BTC spot
# from Kucoin with 30 minute
# intervals
BTC <- try(
  cryptoQuotes::getQuote(
    ticker   = 'BTC-USDT',
    source   = 'kucoin',
    interval = '30m',
    futures  = FALSE,
    from     = Sys.Date() - 1
  )
)

# 2) chart the spot price
# using the chart
# function
if (!inherits(BTC, 'try-error')){

  cryptoQuotes::chart(
    chart = cryptoQuotes::kline(BTC) %>%
      cryptoQuotes::addVolume() %>%
      cryptoQuotes::addBBands()
  )

}

# script end;