Get a quote on a cryptopair from one of the major exchanges

This function returns a crypto quote from one of the available exchanges. The function supports futues and spot markets

Usage

getQuote(
  ticker,
  source = "binance",
  futures = TRUE,
  interval,
  from = NULL,
  to = NULL
)

Arguments

ticker

A character vector of length 1. Uppercase.

source

A character vector of length 1. Currently only binance is supported.

futures

A logical value. TRUE by default. If FALSE, the function will return spot prices.

interval

A character vector of length 1. See availableIntervals() for a full list.

from

A character vector of length 1. Given in %Y-%m-%d format.

to

A character vector of length 1.Given in %Y-%m-%d format.

Examples

# 1) load perpetual
# futures from Binance
# with 15m intervals
perpAtom <- try(
  cryptoQuotes::getQuote(
    ticker  = 'ATOMUSDT',
    source = 'binance',
    interval = '15m',
    futures = TRUE
  )
)
#> Error in response[, 1] : incorrect number of dimensions

# 2) chart the futures
# using the chart
# function
if (!inherits(perpAtom, 'try-error')){
  cryptoQuotes::chart(
    chart = cryptoQuotes::kline(perpAtom) %>%
      cryptoQuotes::addVolume()   %>%
        cryptoQuotes::addBBands(cols = c('Close'))
  )
}

# NOTE: Without the try
# the examples fails on
# Github Actions

# script end;